Move template loader to lib and maya api
