+++ openpype/lib/abstract_load_template.py
